iterators: move trait to lib.rs
